backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-deep-serializer==0.1.3
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
executing==2.1.0
filelock==3.16.1
fonttools==4.49.0
html5lib==1.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
joblib==1.4.0
kiwisolver==1.4.5
lxml==5.3.0
matplotlib==3.8.3
matplotlib-inline==0.1.7
nltk==3.8.1
numpy==1.26.4
oscrypto==1.3.0
packaging==23.2
parso==0.8.4
pillow==10.2.0
pipenv==2024.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyparsing==3.1.1
pypdf==5.1.0
pyreadline==2.1
python-bidi==0.6.3
python-dateutil==2.8.2
PyYAML==6.0.2
qrcode==8.0
regex==2023.12.25
reportlab==4.2.5
requests==2.32.3
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
textblob==0.18.0.post0
tinycss2==1.4.0
tqdm==4.66.2
traitlets==5.14.3
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.28.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
